backend-python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
influxdb-client==1.40.0
paho-mqtt==2.0.0
python-multipart==0.0.6
pymysql==1.1.0
